test: :white_check_mark: add first test

+++ src/randomness-vulnerabilities-1/Game.sol
@@ -10,11 +10,11 @@ pragma solidity ^0.8.13;
 contract Game {
     constructor() payable {}
 
-    function play(uint guess) external {
-        uint number = uint(keccak256(abi.encodePacked(block.timestamp, block.number, block.difficulty)));
+    function play(uint256 guess) external {
+        uint256 number = uint256(keccak256(abi.encodePacked(block.timestamp, block.number, block.difficulty)));
 
         if (guess == number) {
-            (bool sent, ) = msg.sender.call{value: address(this).balance}("");
+            (bool sent,) = msg.sender.call{value: address(this).balance}("");
             require(sent, "Failed to send ETH");
         }
     }
